requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-migrate==4.0.4
flask-cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.10
bcrypt==4.0.1
flask-jwt-extended==4.5.2
Werkzeug==2.3.7
pytz==2024.1